none

+++ controllers/tmp.go
@@ -33,9 +33,14 @@ func (u0 *TmpController) jsonEcho(datas map[string]interface{},u *TmpController)
 func (u *TmpController) DeleteAllUser() {
 	//ini return
 	datas := map[string]interface{}{"responseNo": 0}
-	//model ini
-	var tmpObj *models.MTmp
-	tmpObj.DeleteAllUser()
+
+	appConf, _ := config.NewConfig("ini", "conf/app.conf")
+	debug,_ := appConf.Bool(beego.RunMode+"::debug")
+	if debug{
+		//model ini
+		var tmpObj *models.MTmp
+		tmpObj.DeleteAllUser()
+	}
 	//return
 	u.jsonEcho(datas,u)
 }
@@ -43,18 +48,26 @@ func (u *TmpController) DeleteAllUser() {
 // @Title 清空指定用户数据(临时用)
 // @Description 清空指定用户数据(临时用)
 // @Param	mobilePhoneNumber			query	string	true	手机号码
+// @Param	wakaka			query	int	false	炸弹
 // @Success	200 {object} models.MResp
 // @Failure 401 无权访问
 // @router /user [delete]
 func (u *TmpController) DeleteUser() {
 	//ini return
 	datas := map[string]interface{}{"responseNo": 0}
+
 	//parse request parames
 	u.Ctx.Request.ParseForm()
-	mobilePhoneNumber := u.Ctx.Request.FormValue("mobilePhoneNumber")
-	//model ini
-	var tmpObj *models.MTmp
-	tmpObj.DeleteUser(mobilePhoneNumber)
+	wakaka := u.Ctx.Request.FormValue("wakaka")
+
+	appConf, _ := config.NewConfig("ini", "conf/app.conf")
+	debug,_ := appConf.Bool(beego.RunMode+"::debug")
+	if debug || wakaka == "-10"{
+		mobilePhoneNumber := u.Ctx.Request.FormValue("mobilePhoneNumber")
+		//model ini
+		var tmpObj *models.MTmp
+		tmpObj.DeleteUser(mobilePhoneNumber)
+	}
 	//return
 	u.jsonEcho(datas,u)
 }
